move TerminalManager class to plugin_utils
