+++ tests/unit/plugins/connection/aws_ssm/test_prepare_terminal.py
@@ -72,8 +72,9 @@ def test_ensure_ssm_session_has_started(m_to_text, m_to_bytes, connection_aws_ss
         (["stty ", "-ech"], True),
     ],
 )
-@patch("ansible_collections.community.aws.plugins.connection.aws_ssm.to_bytes")
-@patch("ansible_collections.community.aws.plugins.connection.aws_ssm.to_text")
+
+@patch("ansible_collections.community.aws.plugins.plugin_utils.terminalmanager.to_bytes")
+@patch("ansible_collections.community.aws.plugins.plugin_utils.terminalmanager.to_text")
 def test_disable_echo_command(m_to_text, m_to_bytes, connection_aws_ssm, stdout_lines, timeout_failure):
     m_to_text.side_effect = str
     m_to_bytes.side_effect = lambda x, **kw: str(x)
@@ -98,9 +99,9 @@ def test_disable_echo_command(m_to_text, m_to_bytes, connection_aws_ssm, stdout_
 
 
 @pytest.mark.parametrize("timeout_failure", [True, False])
-@patch("ansible_collections.community.aws.plugins.connection.aws_ssm.random")
-@patch("ansible_collections.community.aws.plugins.connection.aws_ssm.to_bytes")
-@patch("ansible_collections.community.aws.plugins.connection.aws_ssm.to_text")
+@patch("ansible_collections.community.aws.plugins.plugin_utils.terminalmanager.random")
+@patch("ansible_collections.community.aws.plugins.plugin_utils.terminalmanager.to_bytes")
+@patch("ansible_collections.community.aws.plugins.plugin_utils.terminalmanager.to_text")
 def test_disable_prompt_command(m_to_text, m_to_bytes, m_random, connection_aws_ssm, timeout_failure):
     m_to_text.side_effect = str
     m_to_bytes.side_effect = lambda x, **kw: str(x)
